refactor:rename package name to racingcar
