+++ src/main/java/racingcar/InputView.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar;
 
 import java.util.List;
