+++ src/main/java/racingcar/MovementCondition.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar;
 
 public interface MovementCondition {
     boolean isMovable();
