+++ src/main/java/racingcar/Observer.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar;
 
 public interface Observer<T> {
     void observe(T event);
