+++ src/main/java/racingcar/OutputView.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar;
 
 import java.util.List;
