+++ src/main/java/racingcar/RacingCarMain.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar;
 
 import java.util.ArrayList;
 import java.util.List;
