+++ src/main/java/racingcar/RacingSubject.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar;
 
 import java.util.HashSet;
 import java.util.Set;
