+++ src/main/java/racingcar/RandomMovementCondition.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar;
 
 import java.util.Random;
