+++ src/main/java/racingcar/Subject.java
@@ -1,4 +1,4 @@
-package RacingCar;
+package racingcar;
 
 public interface Subject<T> {
     void addObserver(Observer<T> observer);
