+++ src/test/java/RacingCarTest/RacingTest.java
@@ -1,14 +1,12 @@
 package RacingCarTest;
 
-import RacingCar.*;
-import org.junit.jupiter.api.Disabled;
+import racingcar.*;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
 import java.util.ArrayList;
-import java.util.List;
 import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.*;
